Force review of query quantity and adjust handling for missing products_attributes_id
Discovered that there was at least a system in existence that cached results
returning a result that the query was actually done processing even though
there were records to be processed in the review.  Therefore, changed to
a more active review (without requiring the query to be asked again) such
that if there are results, then capture those results, if not then don't
use the value to support display.  If the value is found but the option
and the value (both dependent on text) are empty, then remove the
attribute option from the SBA table.  The thought also is that in combining
data from the various tables, that the options_name and options_values_name
would not have results to provide and that they would be empty.

+++ admin/includes/classes/products_with_attributes_stock.php
@@ -243,7 +243,7 @@ class products_with_attributes_stock extends base
       $attributes = $db->Execute($query);
 
       $attributes_output = false;
-      if (!$attributes->EOF) {
+      if ($attributes->RecordCount() > 0) {
         $attributes_output = array(
                                    'option' => $attributes->fields['products_options_name'],
                                    'value' => $attributes->fields['products_options_values_name'],
@@ -480,6 +480,7 @@ class products_with_attributes_stock extends base
                   foreach ($attributes_of_stock as $attri_id)
                   {
                       $stock_attribute = $this->get_attributes_name($attri_id['products_attributes_id'], $attri_id['language_id']/*$_SESSION['languages_id']*/);
+                      if ($stock_attribute === false) continue; // If the products_attributes_id is not found in the selected language then move on.
                       if ($stock_attribute['option'] == '' && $stock_attribute['value'] == '') {
                         // delete stock attribute
                         $db->Execute("DELETE FROM " . TABLE_PRODUCTS_WITH_ATTRIBUTES_STOCK . " WHERE stock_id = " . $attribute_products->fields['stock_id'] . " LIMIT 1;");
